Stop doing Initalization as module
Now that we have a base class, just use it more directly.
