+++ shoes-core/lib/shoes/dsl.rb
@@ -81,7 +81,6 @@ require 'shoes/common/changeable'
 require 'shoes/common/clickable'
 require 'shoes/common/fill'
 require 'shoes/common/hover'
-require 'shoes/common/initialization'
 require 'shoes/common/link_finder'
 require 'shoes/common/positioning'
 require 'shoes/common/remove'
